procfs: get PID from Ubuntu namespace
This patch allows code running on the Android side to properly lookup a
processes PID from the Ubuntu namespace, so that it both sides can refer
to the same process.

This is a short-term solution, however it unblocks a bug which prevents
applications from being launched from the command-line without manual
intervention to get the app to display.